docs: run samples on emulator
Runs samples by default on the emulator. This makes the samples
completely standalone. No setup or configuration is needed in order to
run the samples. Instead, the SampleRunner will automatically download
and start an instance of the emulator, create the sample database, and
then run the sample on the emulator.