docs: add sample for using query hints
Adds a sample for how to use query hints with Cloud Spanner in Entity Framework Core.

The sample tags a query with a specific text to indicate that the query should be intercepted
by a command interceptor that will add a query hint to the query.

This sample will only work once https://github.com/googleapis/google-cloud-dotnet/pull/6848
has been merged and released, as the client library currently does not support untyped commands
that start with either a comment or a statement hint.