fix: dispose the underlying SpannerDataReader
The SpannerDataReaderWithChecksum did not dispose the underlying SpannerDataReader
when it was closed/disposed. This caused unnecessary logging to happen that indicated
that a session could have been leaked. The failure to close the underlying reader
would however not cause a session leak, as the session is managed by the surrounding
transaction, which again will return the session to the pool when the transaction is
disposed.