feat: support THEN RETURN clauses
Use THEN RETURN for returning data from INSERT/UPDATE/DELETE statements.
This reduces the number of round-trips to Spanner for most use-cases.
THEN RETURN clauses are not supported in BatchDml, which means that
multiple DML statements that need to return values are translated to
separate DML statements, instead of a BatchDmlRequest followed by N
SELECT statements. Even in those cases, using THEN RETURN reduces the
number of round-trips by 1.

The fact that Batch DML does not support THEN RETURN is the main reason
that it is recommended to use client-side generated UUIDs for primary
keys when using Entity Framework. This will allow DML statements to be
batched.