feat: retry DML with THEN RETURN in transactions
DML statements with a THEN RETURN clause need to be executed using the
ExecuteReader method. These were normally not executed in a retry loop,
as they were up until now only used for queries. Queries are executed
during the first call to Reader#Read(), which meant that a retry loop
was not needed for calls to ExecuteReader. DML statements with a THEN
RETURN clause are executed directly, and therefore need a retry loop.

This fix adds a retry loop for that, and tests for DML statements with
a THEN RETURN clause.