perf: translate Contains to a parameterized IN query (#472)
IN queries were automatically translated by EF Core to queries that
used `IN (value1, value2, ..., valueN)`. This requires Spanner to
re-compile the query for every possible permutation of the list of
values.

This change registers a specific SpannerInExpression that is used
when a query using a fixed list of values is being generated. These
queries are then translated to queries of this form:

`IN UNNEST (@values)`

The `@values` parameter is of type ARRAY.