perf: use inline-begin-transaction
The Spanner client now uses inline-begin-transaction. This means that the client
includes a BeginTransaction option with the first command in a transaction, instead
of executing a separate BeginTransaction RPC. This improves performance for most
workloads, and reduces the probability of getting a TransactionOutcomeUnknown error.

Transactions that only write mutations (blind writes) do not benefit from this
optimization, as these transactions do not execute any SQL commands that can include
the BeginTransaction option. For these transactions, the Spanner client will still
execute a separate BeginTransaction RPC.