fix: identity options in annotations should be serialized
Identity options that are included in annotations for columns when creating a new
migration should be serialized to a string, so these can be automatically written
to the generated code. This prevents errors like "Cannot scaffold C# literals of
type 'Google.Cloud.EntityFrameworkCore.Spanner.Metadata.SpannerIdentityOptionsData'"
from occurring when creating a new migration.